+++ CHANGELOG.md
@@ -3,9 +3,13 @@
 ## [2.4.0] - TBD
 #### Feature
 - Change venue when match warmup starts
+- New endpoint `/tournament/<id>/matches/result` returning results for all matches in tournament
 
 #### Changed
 - Switched from string to time for datetimes to correctly support timezones
+
+#### Fixed
+- Correctly rematch for matches with Bots
 
 ## [2.3.0] - 2022-03-06
 #### Feature